Removes domain classes and refactorizes the code
The AssignedGamesSession entity class has been moved to the MTC Games
project. As this was the only entity class, the domain folder was
removed.

In addition, the code has been refactorized in several parts: parameter
type checking was added to some methods, components was divided into
several folders and internal utility methods have been renamed and
marked as private in GamesSessionEndpoint class.