Adds support for open/closed sessions
An open session is a session that can be played, because the current day
is between start and end games session dates. This commit adds a new
behaviour to session list to show different states for open and closed
games sessions. This way, closed sessions can't be played until they are
open.