Fixes some login errors
Assigned sessions were currently loaded before user's login, causing
errors when the username was not already stored in local storage.

In addition, the local storage keys used were the same as the used in
other MTC applications. This made that, when several applications are
deployed in the same domain, credentials were overwritten by/to other
applications.